pyproject.toml
[tool.poetry]
name = "craic"
version = "0.1.0"
description = "A python package for identifing crystal phases during nucleation with the help of artificial intelligence classifiers"
authors = ["aksam432 <akshaykrishnaak9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.2.2"
pybind11 = "^2.10.4"
pyscal2 = {git = "https://github.com/pyscal/pyscal"}
pyyaml = "^6.0"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
